backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
sqlalchemy==1.4.23
pillow==8.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.8.2
email-validator==1.1.3
aiosqlite==0.19.0
pandas==2.1.4
python-magic==0.4.27
opencv-python==4.8.1.78
python-dotenv==1.0.0
aiofiles==0.8.0 